requirements.txt
flask[async]==3.0.2
python-kasa==0.6.2.1